Account for Forked PRs + refactor testing (#45)
## PR Overview

This PR includes necessary updated for getting the merge action to work
for forked PRs. When this action runs on `pull_request` events for
forked PRs, the action will not have access to any secrets, most likely
including the trunk token necessary for this action to work. To account
for this, the `setImpactedTargets` API allows for uploading impacted
targets using the github workflow run ID of the job that is uploading
impacted targets. If the provided ID belongs to an in-progress run from
a forked PR, and the SHA within the request matches that SHA of that
workflow run ID, then the upload will be allowed

This PR handles the above, making it so that if the workflow is coming
from a fork, we properly make the `setImpactedTargets` API call using
the forked workflow run ID

## Changes to how this action is used

None

## Testing

I added tests verifying the new conditions around not always needing the
API token. Additionally, I refactored the tests to make future changes
like this easier

## Other changes

* In adjusting testing, I noticed some bugs within the bash script and
fixed those as well
* I updated trunk